Build two versions.
