+++ demo/index.html
@@ -11,11 +11,7 @@
     }
 
     button {
-      position: absolute;
-      top: 50%;
-      left: 50%;
-      margin-top: -24px;
-      margin-left: -125px;
+      margin: 50px;
     }
   </style>
 </head>
