+++ demo/index.js
@@ -2,6 +2,7 @@ import React, { Component } from 'react';
 import ReactDOM from 'react-dom';
 import { Router, Route, Link, browserHistory } from 'react-router';
 import FacebookLogin from '../src/facebook';
+import FacebookLoginWithButton from '../src/facebook-with-button'
 
 const responseFacebook = (response) => {
   console.log(response);
@@ -12,12 +13,28 @@ class Base extends Component {
     return (
       <div>
         <Link to="/dummy">Route to dummy page</Link>
-        <FacebookLogin
-          appId="1088597931155576"
-          autoLoad
-          callback={responseFacebook}
-          icon="fa-facebook"
-        />
+
+        <div>
+          <p>Facebook login with default button and styling</p>
+          <FacebookLoginWithButton
+            appId="1088597931155576"
+            autoLoad
+            callback={responseFacebook}
+            icon="fa-facebook"
+          />
+        </div>
+
+        <div>
+          <p>Facebook login with render prop (and no styling provided out the box)</p>
+          <FacebookLogin
+            appId="1088597931155576"
+            autoLoad
+            callback={responseFacebook}
+            render={renderProps => (
+              <button onClick={renderProps.onClick}>This is my custom FB button</button>
+            )}
+          />
+        </div>
       </div>
     );
   }
