+++ src/facebook.js
@@ -1,7 +1,6 @@
 // @flow
 import React from 'react';
 import PropTypes from 'prop-types';
-import styles from '../styles/facebook.scss';
 import getParamsFromObject from './objectToParams';
 
 const getIsMobile = () => {
@@ -16,17 +15,6 @@ const getIsMobile = () => {
   return isMobile;
 };
 
-// https://www.w3.org/TR/html5/disabled-elements.html#disabled-elements
-const _shouldAddDisabledProp = (tag) => [
-  'button',
-  'input',
-  'select',
-  'textarea',
-  'optgroup',
-  'option',
-  'fieldset',
-].indexOf((tag + '').toLowerCase()) >= 0;
-
 class FacebookLogin extends React.Component {
 
   static propTypes = {
@@ -39,42 +27,29 @@ class FacebookLogin extends React.Component {
     scope: PropTypes.string,
     returnScopes: PropTypes.bool,
     redirectUri: PropTypes.string,
-    textButton: PropTypes.string,
-    typeButton: PropTypes.string,
     autoLoad: PropTypes.bool,
     disableMobileRedirect: PropTypes.bool,
     isMobile: PropTypes.bool,
-    size: PropTypes.string,
     fields: PropTypes.string,
-    cssClass: PropTypes.string,
     version: PropTypes.string,
-    icon: PropTypes.any,
     language: PropTypes.string,
     onClick: PropTypes.func,
-    containerStyle: PropTypes.object,
-    buttonStyle: PropTypes.object,
-    children: PropTypes.node,
-    tag: PropTypes.oneOfType([PropTypes.node, PropTypes.func]),
     onFailure: PropTypes.func,
+    render: PropTypes.func.isRequired,
   };
 
   static defaultProps = {
-    textButton: 'Login with Facebook',
-    typeButton: 'button',
     redirectUri: typeof window !== 'undefined' ? window.location.href : '/',
     scope: 'public_profile,email',
     returnScopes: false,
     xfbml: false,
     cookie: false,
     reAuthenticate: false,
-    size: 'metro',
     fields: 'name',
-    cssClass: 'kep-login-facebook',
     version: '2.3',
     language: 'en_US',
     disableMobileRedirect: false,
     isMobile: getIsMobile(),
-    tag: 'button',
     onFailure: null,
   };
 
@@ -208,59 +183,20 @@ class FacebookLogin extends React.Component {
     }
   };
 
-  style() {
-    const defaultCSS = this.constructor.defaultProps.cssClass;
-    if (this.props.cssClass === defaultCSS) {
-      return <style dangerouslySetInnerHTML={{ __html: styles }}></style>;
-    }
-    return false;
-  }
-
-  // [AdGo] 20.11.2016 - coult not get container class to work
-  containerStyle() {
-    const style = { transition: 'opacity 0.5s' };
-    if (this.state.isProcessing || !this.state.isSdkLoaded || this.props.isDisabled) {
-      style.opacity = 0.6;
-    }
-    return Object.assign(style, this.props.containerStyle);
-  }
-
-  renderOwnButton() {
-    const { cssClass, size, icon, textButton, typeButton, buttonStyle } = this.props;
-    const isIconString = typeof icon === 'string';
-    const optionalProps = {};
-    if (this.props.isDisabled && _shouldAddDisabledProp(this.props.tag)) {
-      optionalProps.disabled = true;
-    }
-    return (
-      <span style={ this.containerStyle() }>
-        {isIconString && (
-          <link
-            rel="stylesheet"
-            href="//maxcdn.bootstrapcdn.com/font-awesome/4.5.0/css/font-awesome.min.css"
-          />
-        )}
-        <this.props.tag
-          type={typeButton}
-          className={`${cssClass} ${size}`}
-          style={ buttonStyle }
-          onClick={this.click}
-          {...optionalProps}
-        >
-          {icon && isIconString && (
-            <i className={`fa ${icon}`}></i>
-          )}
-          {icon && !isIconString && icon}
-          {textButton}
-        </this.props.tag>
-        {this.style()}
-      </span>
-    );
-  }
-
   render() {
-    const { children } = this.props;
-    return children ? <span onClick={this.click}>{ children }</span> : this.renderOwnButton();
+    const { render } = this.props
+
+    if (render) {
+      const propsForRender = {
+        onClick: this.click,
+        isDisabled: !!this.props.isDisabled,
+        isProcessing: this.state.isProcessing,
+        isSdkLoaded: this.state.isSdkLoaded,
+      }
+      return this.props.render(propsForRender)
+    } else {
+      throw new Error('ReactFacebookLogin requires a render prop to render')
+    }
   }
 }
